chore: Avoid implicit cast in math operations
Resolve Sonar issues of type java:S2184 - Math operands should be cast before assignment

related to #88

+++ qa/performance-tests-engine/src/main/java/org/operaton/bpm/qa/performance/engine/framework/activitylog/ActivityCountAggregator.java
@@ -175,7 +175,7 @@ public class ActivityCountAggregator extends TabularResultAggregator {
   }
 
   protected int calculateTimeSlot(Date date, long firstTimestamp) {
-    return Math.round((date.getTime() - firstTimestamp) / INTERVAL);
+    return Math.round((float) (date.getTime() - firstTimestamp) / INTERVAL);
   }
 
   static class ActivityCount {
